+++ infrastructure/src/main/resources/db/migration/common/V20230414.1501__create_table.sql
@@ -1,6 +1,7 @@
-create table if not exists travelbuddy.flightspecial 
+create table if not exists travelbuddy.flightspecial
 (
-    id                              int4 GENERATED ALWAYS AS IDENTITY,
+--    id                              int4 GENERATED ALWAYS AS IDENTITY,
+    id                              int8 GENERATED ALWAYS AS IDENTITY,
     header                          varchar(255) not null,
     body                            varchar(255),
     origin                          varchar(255),
